requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
WTForms==3.1.2
Flask-Login==0.6.3
Flask-Migrate==4.0.7
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9

# HTTP-клиент для интеграции с Go-сервисом:
requests==2.32.3

# Валидация email в WTForms:
email-validator==2.2.0
requests==2.32.3
email-validator==2.2.0
python-dotenv==1.0.1
